generalize hermitian_conjugated and move to utils
