+++ src/openfermion/measurements/_equality_constraint_projection.py
@@ -14,8 +14,8 @@
 import numpy
 import scipy
 
-from openfermion.ops import FermionOperator, hermitian_conjugated
-from openfermion.utils import count_qubits
+from openfermion.ops import FermionOperator
+from openfermion.utils import count_qubits, hermitian_conjugated
 
 from ._rdm_equality_constraints import two_body_fermion_constraints
