+++ src/openfermion/ops/__init__.py
@@ -15,7 +15,6 @@ from ._code_operator import (BinaryCode,
                              linearize_decoder)
 from ._symbolic_operator import SymbolicOperator
 from ._fermion_operator import (FermionOperator,
-                                hermitian_conjugated,
                                 normal_ordered)
 from ._qubit_operator import QubitOperator
 from ._polynomial_tensor import general_basis_change, PolynomialTensor
