+++ src/openfermion/utils/__init__.py
@@ -20,8 +20,9 @@ from ._commutators import anticommutator, commutator, double_commutator
 from ._grid import Grid
 
 from ._operator_utils import (count_qubits, eigenspectrum, fourier_transform,
-                              get_file_path, inverse_fourier_transform,
-                              is_identity, load_operator, save_operator)
+                              get_file_path, hermitian_conjugated,
+                              inverse_fourier_transform, is_identity,
+                              load_operator, save_operator)
 
 from ._slater_determinants import (gaussian_state_preparation_circuit,
                                    slater_determinant_preparation_circuit)
