+++ src/openfermion/utils/_operator_utils.py
@@ -31,6 +31,23 @@ class OperatorUtilsError(Exception):
     pass
 
 
+def hermitian_conjugated(operator):
+    """Return Hermitian conjugate of operator."""
+    if isinstance(operator, FermionOperator):
+        conjugate_operator = FermionOperator()
+        for term, coefficient in iteritems(operator.terms):
+            conjugate_term = tuple([(tensor_factor, 1 - action) for
+                                    (tensor_factor, action) in reversed(term)])
+            conjugate_operator.terms[conjugate_term] = numpy.conjugate(
+                    coefficient)
+    elif isinstance(operator, QubitOperator):
+        conjugate_operator = QubitOperator()
+        for term, coefficient in operator.terms.items():
+            conjugate_operator.terms[term] = coefficient.conjugate()
+
+    return conjugate_operator
+
+
 def count_qubits(operator):
     """Compute the minimum number of qubits on which operator acts.
